Add files via upload
The "Login" and "Signup" buttons now route to different pages within the project. When "Signup" is clicked, the sign up page is presented to the user. An additional text box was added for the user to confirm the password and alert() dialogs are presented whenever fields aren't filled or account types are not selected. Page routing is handled within test.js, presenting the html page based on account type once authenticated.